% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeliveryLocation.R
\name{DeliveryLocationUI}
\alias{DeliveryLocationUI}
\title{交货地点生成器界面}
\usage{
DeliveryLocationUI(
  tabTitle = "Delivery Location",
  colTitles = c("Filter Area", "Operation Area", "Result Area"),
  widthRates = c(6, 6, 12),
  func_left = DeliveryLocationUI_left,
  func_right = DeliveryLocationUI_right,
  func_bottom = DeliveryLocationUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
交货地点生成器界面
}
\examples{
DeliveryLocationUI()
}
